crud/Authentification

+++ app/src/main/java/com/example/proform/listemp.java
@@ -1,5 +1,6 @@
 package com.example.proform;
 
+import android.app.ProgressDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Canvas;
@@ -8,8 +9,11 @@ import android.graphics.Paint;
 import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
+import android.text.InputType;
+import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
+import android.widget.EditText;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
@@ -34,6 +38,7 @@ import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
 
 import java.util.ArrayList;
@@ -45,6 +50,7 @@ public class listemp extends AppCompatActivity {
     private UserAdapter adapter;
     private List<User> userList;
     private FirebaseAuth mAuth;
+    private boolean testD;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -52,7 +58,7 @@ public class listemp extends AppCompatActivity {
         setContentView(R.layout.activity_listemp);
 
         mAuth = FirebaseAuth.getInstance();
-
+         testD=false;
         recyclerView = findViewById(R.id.recyclerView);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
 
@@ -76,14 +82,24 @@ public class listemp extends AppCompatActivity {
                         } else {
                             user.setAdmin(false);
                         }
-                        if (!user.isAdmin()) {
-                            userList.add(user);
+
+                        FirebaseUser currentUser = mAuth.getCurrentUser();
+                        if (currentUser != null && user.getPoste().equals("Chef personnelle") && currentUser.getUid().equals(user.getUserId())) {
+                            if (user.getPoste().equals("Transporter")) {
+                                userList.add(user);
+                            }
+                        } else {
+                            if (!user.isAdmin()) {
+                                userList.add(user);
+                            }
                         }
                     }
                 }
                 Log.d("UserListDebug", "UserList size: " + userList.size());
                 adapter.notifyDataSetChanged();
             }
+
+
 
             @Override
             public void onCancelled(@NonNull DatabaseError databaseError) {
@@ -122,9 +138,6 @@ public class listemp extends AppCompatActivity {
                     if (direction == ItemTouchHelper.LEFT) {
                         showDeleteConfirmationDialog(position);
                     } else if (direction == ItemTouchHelper.RIGHT) {
-                        Intent intent = new Intent(listemp.this, updatep.class);
-                        intent.putExtra("user", userList.get(position));
-                        startActivity(intent);
                         adapter.notifyItemChanged(position);
                     }
                 } else {
@@ -158,44 +171,31 @@ public class listemp extends AppCompatActivity {
                 super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
                 return;
             }
-
-            // Draw the appropriate background based on swipe direction
             if (dX > 0) {
-                // Draw the gray background for update
                 backgroundU.setBounds(itemView.getLeft(), itemView.getTop(), itemView.getLeft() + (int) dX, itemView.getBottom());
                 backgroundU.draw(c);
-
-                // Calculate position of update icon
                 int updateIconTop = itemView.getTop() + (itemHeight - updateIcon.getIntrinsicHeight()) / 2;
                 int updateIconMargin = (itemHeight - updateIcon.getIntrinsicHeight()) / 2;
                 int updateIconLeft = itemView.getLeft() + updateIconMargin;
                 int updateIconRight = itemView.getLeft() + updateIconMargin + updateIcon.getIntrinsicWidth();
                 int updateIconBottom = updateIconTop + updateIcon.getIntrinsicHeight();
-
-                // Draw the update icon
                 updateIcon.setBounds(updateIconLeft, updateIconTop, updateIconRight, updateIconBottom);
                 updateIcon.draw(c);
             } else {
-                // Draw the red background for delete
                 background.setBounds(itemView.getRight() + (int) dX, itemView.getTop(), itemView.getRight(), itemView.getBottom());
                 background.draw(c);
 
-                // Calculate position of delete icon
                 int deleteIconTop = itemView.getTop() + (itemHeight - deleteIcon.getIntrinsicHeight()) / 2;
                 int deleteIconMargin = (itemHeight - deleteIcon.getIntrinsicHeight()) / 2;
                 int deleteIconLeft = itemView.getRight() - deleteIconMargin - deleteIcon.getIntrinsicWidth();
                 int deleteIconRight = itemView.getRight() - deleteIconMargin;
                 int deleteIconBottom = deleteIconTop + deleteIcon.getIntrinsicHeight();
-
-                // Draw the delete icon
                 deleteIcon.setBounds(deleteIconLeft, deleteIconTop, deleteIconRight, deleteIconBottom);
                 deleteIcon.draw(c);
             }
 
             super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
         }
-
-
         private void clearCanvas(Canvas c, Float left, Float top, Float right, Float bottom) {
             Paint paint = new Paint();
             paint.setColor(((ColorDrawable) background).getColor());
@@ -203,22 +203,60 @@ public class listemp extends AppCompatActivity {
         }
 
         private void showDeleteConfirmationDialog(final int position) {
-            AlertDialog.Builder builder = new AlertDialog.Builder(listemp.this);
-            builder.setMessage("Are you sure you want to delete this item?")
-                    .setPositiveButton("Delete", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            deleteUser(position);
-                        }
-                    })
-                    .setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            adapter.notifyItemChanged(position);
-                        }
-                    })
-                    .show();
+            final EditText input = new EditText(listemp.this);
+            input.setInputType(InputType.TYPE_CLASS_TEXT | InputType.TYPE_TEXT_VARIATION_PASSWORD);
+
+            // Check if testD is false before showing the dialog
+            if (!testD) {
+                AlertDialog.Builder builder = new AlertDialog.Builder(listemp.this);
+                builder.setView(input);
+                builder.setMessage("Enter Admin Password:")
+                        .setPositiveButton("Delete", new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface dialog, int which) {
+                                String adminPassword = input.getText().toString().trim();
+                                if (!TextUtils.isEmpty(adminPassword)) {
+                                    DatabaseReference usersRef = FirebaseDatabase.getInstance().getReference("users");
+                                    Query adminQuery = usersRef.orderByChild("poste").equalTo("Admin").limitToFirst(1);
+                                    adminQuery.addListenerForSingleValueEvent(new ValueEventListener() {
+                                        @Override
+                                        public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                                            for (DataSnapshot userSnapshot : dataSnapshot.getChildren()) {
+                                                String adminPasswordFromDB = userSnapshot.child("password").getValue(String.class);
+                                                if (adminPassword.equals(adminPasswordFromDB)) {
+                                                    deleteUser(position);
+                                                    return;
+                                                }
+                                            }
+                                            Toast.makeText(listemp.this, "Incorrect Admin Password", Toast.LENGTH_SHORT).show();
+                                            adapter.notifyItemChanged(position);
+                                        }
+
+                                        @Override
+                                        public void onCancelled(@NonNull DatabaseError databaseError) {
+                                            Toast.makeText(listemp.this, "Failed to retrieve Admin Password", Toast.LENGTH_SHORT).show();
+                                            adapter.notifyItemChanged(position);
+                                        }
+                                    });
+                                } else {
+                                    Toast.makeText(listemp.this, "Please enter Admin Password", Toast.LENGTH_SHORT).show();
+                                    adapter.notifyItemChanged(position);
+                                }
+                            }
+                        })
+                        .setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface dialog, int which) {
+                                adapter.notifyItemChanged(position);
+                            }
+                        })
+                        .show();
+                testD = true;
+            }
         }
+
+
+
 
 
 
@@ -226,55 +264,42 @@ public class listemp extends AppCompatActivity {
             final DatabaseReference databaseReference = FirebaseDatabase.getInstance().getReference("users");
             final String userId = userList.get(position).getUserId();
             Log.d("UserID", userId);
-
+            ProgressDialog progressDialog = ProgressDialog.show(listemp.this, "", "Please wait...", true);
             FirebaseUser currentUser = FirebaseAuth.getInstance().getCurrentUser();
-                // Sign in with the credentials of the user being deleted
-                FirebaseAuth.getInstance().signInWithEmailAndPassword(userList.get(position).getEmail(), userList.get(position).getPassword())
-                        .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
-                            @Override
-                            public void onComplete(@NonNull Task<AuthResult> task) {
-                                if (task.isSuccessful()) {
-                                    FirebaseUser firebaseUser = FirebaseAuth.getInstance().getCurrentUser();
-                                    if (firebaseUser != null) {
-                                        firebaseUser.delete().addOnCompleteListener(new OnCompleteListener<Void>() {
-                                            @Override
-                                            public void onComplete(@NonNull Task<Void> task) {
-                                                if (task.isSuccessful()) {
-                                                    // User authentication and deletion succeeded, now delete user data from database
-                                                    databaseReference.child(userId).removeValue(new DatabaseReference.CompletionListener() {
-                                                        @Override
-                                                        public void onComplete(@Nullable DatabaseError error, @NonNull DatabaseReference ref) {
+            FirebaseAuth.getInstance().signInWithEmailAndPassword(userList.get(position).getEmail(), userList.get(position).getPassword())
+                    .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                        @Override
+                        public void onComplete(@NonNull Task<AuthResult> task) {
+                            if (task.isSuccessful()) {
+                                FirebaseUser firebaseUser = FirebaseAuth.getInstance().getCurrentUser();
+                                if (firebaseUser != null) {
+                                    firebaseUser.delete().addOnCompleteListener(new OnCompleteListener<Void>() {
+                                        @Override
+                                        public void onComplete(@NonNull Task<Void> task) {
+                                            if (task.isSuccessful()) {
+                                                databaseReference.child(userId).removeValue(new DatabaseReference.CompletionListener() {
+                                                    @Override
+                                                    public void onComplete(@Nullable DatabaseError error, @NonNull DatabaseReference ref) {
+                                                        progressDialog.dismiss();
                                                             if (error == null) {
-                                                                userList.remove(position);
-                                                                adapter.notifyItemRemoved(position);
-                                                                Toast.makeText(listemp.this, "User deleted", Toast.LENGTH_SHORT).show();
-                                                                finish();
-                                                            } else {
-                                                                // Error deleting user data from database
-                                                                Toast.makeText(listemp.this, "Failed to delete user from Realtime Database: " + error.getMessage(), Toast.LENGTH_SHORT).show();
+                                                            if (position >= 0 && position < userList.size()) {
+                                                            userList.remove(position);
+                                                             adapter.notifyItemRemoved(position);
+                                                             testD=true;
+                                                            Toast.makeText(listemp.this, "User deleted", Toast.LENGTH_SHORT).show();
                                                             }
                                                         }
-                                                    });
-                                                } else {
-                                                    // Error deleting user from Firebase Authentication
-                                                    Toast.makeText(listemp.this, "Failed to delete user from Firebase Authentication: " + task.getException().getMessage(), Toast.LENGTH_SHORT).show();
-                                                }
+                                                    }
+                                                });
                                             }
-                                        });
-                                    }
-                                } else {
-                                    // Failed to authenticate user
-                                    Toast.makeText(listemp.this, "Failed to authenticate user: " + task.getException().getMessage(), Toast.LENGTH_SHORT).show();
+                                        }
+                                    });
                                 }
+                            } else {
+                                Toast.makeText(listemp.this, "Failed to authenticate user: " + task.getException().getMessage(), Toast.LENGTH_SHORT).show();
                             }
-                        });
-            }
+                        }
+                    });
         }
-
-
-
-
     }
-
-
-
+}
